app/app/main.py
# main.py
from fastapi import FastAPI
from starlette.middleware.cors import CORSMiddleware

from app.api import user, word

app = FastAPI()

#: Configure CORS
origins = [
    "http://localhost:8080",
]

app.add_middleware(
    CORSMiddleware,
    allow_origins=origins,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(user.router)
app.include_router(word.router)